tests: introduce preprocessed test cases
runtests now runs a variable replacement on the source file before the
test is to be executed and saves the updated version in a temporary file
(log/test[num]) so that all test case readers can get that (if present)
and thus enjoy the powers of proper test case variable substitution.

This is important to allow complete port number freedom.

Test 309 is now updated to work with a non-fixed port number thanks to this.